+++ oilpriceapi/__init__.py
@@ -25,6 +25,7 @@ from oilpriceapi.exceptions import (
     PermissionDeniedError,
     RateLimitError,
     ServerError,
+    SubscriptionIntervalError,
     TimeoutError,
     ValidationError,
 )
@@ -63,6 +64,7 @@ __all__ = [
     "DataNotFoundError",
     "ServerError",
     "FuturesContractError",
+    "SubscriptionIntervalError",
     "ValidationError",
     "NetworkError",
     "TimeoutError",